docker-images/slack-prs/requirements.txt
requests==2.8.1
websocket-client==0.29.0
iso8601==0.1.10
